+++ nodes/Autotask/helpers/options/index.ts
@@ -1,10 +1,12 @@
 import type { ILoadOptionsFunctions, INodePropertyOptions } from 'n8n-workflow';
-import { getFieldOptions, getFilterFieldOptions, getSortFieldOptions } from './entityOptions';
+import { getFieldOptions, getFilterFieldOptions, getSortFieldOptions, getQueryableEntities, getEntityFields } from './entityOptions';
 
 export {
 	getFieldOptions,
 	getFilterFieldOptions,
 	getSortFieldOptions,
+	getQueryableEntities,
+	getEntityFields,
 };
 
 export type LoadOptionsFunction = (this: ILoadOptionsFunctions) => Promise<INodePropertyOptions[]>;